New transducers, a new function to create paths to the cache
And I figured out the cache actually wasn’t working at all!
That is because the compose_path function wasn’t being used
inside the transducers. So they got saved somewhere else.

It still isn’t working but we should be able to get it to work now.

btw I deleted the module method to transduce because it was super
confusing to have to maintain the same method twice