Repositories now returned in order of last updated.
Using an ordered dictionary for the GitCollection,
a backported implementation of the Python 2.7 data structure.

Also an extra check on initialisation of the git modules:
it tries to reach the head commit. If this fails, the repo is not added.
This catches both empty and corrupted repos.

Sorting is probably heavy, but it happens on __init__ of the class,
not when requesting the list through .get_names()