The gEvent server and why we need it:
If we want to test both the osp-site and the git-api locally at the same
time, the normal Django development server doesn’t cut it: since it is
single-threaded, the server can’t send an api request to itself.

Pierre provides us with a forking server. The only thing that needs to be ironed
out, is that it doesn’t serve the static files from 'public' under '/static/*'
nor does it serve the media files from 'visual' under '/visual/*'