Icebergs listing serverside
Checking the iceberg in the OSP-site through api calls for each repo resulted
in 90 http calls for the home page, and 20+ seconds load times in development.
Therefore an iceberg listing is included in the API now.
It is not super general, but for now the best way to go IMHO.