October 2013, during the 3rd edition of the Typojanchy biennal, OSP drove a workshop of 5 days in Seoul.
Rich from its past adventures of Dingbats Liberation Fest and
accompanied by one of its myth, Fonzie.
4 osp and 10 participants dove into the OCR eye.
Back 3 months ago, they had the opportunity to keep in touch with the
team on site and to practice their pancakes with leeks recipe.
Live  from the SIN salon, they invite you to visit the entrails of
Tesseract  and to take Fonzie through a trip in the images of the
travel.

Open source karaoke flavour.

Welcome!