Make the url for visual previews predictable, drop the cache+database, less code, quicker site!
Attempting to view this url:

http://osp.kitchen/visual/osp.foundry.le-patin-helvete/32cb6d3b71a7b03ed985905d615d9bbb875a1e93/width..600/SoundsOfEurope.png.png

Will make visual culture find the corresponding blob in the git, and render a preview at 600px wide.
The image is returned as the response.

However, the image is also saved to disk, in the folder:
visual/osp.foundry.le-patin-helvete/32cb6d3b71a7b03ed985905d615d9bbb875a1e93/width..600/SoundsOfEurope.png.png

Because this is the same path as the url, a properly configured web-server can first
check if the static file exists, and only then pass on the request to the python app.
By not passing through python, the files should be served quicker and put less strain
on the few python processes available (we had unicorn timeouts).
Finally, the API used to keep track of which previews were already rendered through a database—
that part is now cut out.

On the front end, because these urls are predictable, we can generate them in the django templates:
  uptil now we were asking the url for images through a javascript api after page load.
  this meant a http request for the api, and only then a request for the image
  directly requesting images, we cut http requests in half.

important:  remove 'vc_cache', from INSTALLED_APPS in your settings.py