download ZIP on all projects!
Super easy hack, suggested by Eric, when we were discussing the logic to reproduce the functionnalities of the foundry. Simply put, we download from Github.
this can be seen as yet an other strategy to maintain this relation with our github mirrors.
Just for kicks, the link is super simple to interpret: <a href="https://github.com/osp/osp.{{ repo.category }}.{{ repo.name }}/archive/master.zip">download ZIP</a>