This displays the menu items as blocks in situations of screens under 512 px
I hope this doesn't conflict with @latsami 's 39fc76d95aa34f00d85365b3adde8ddf6efcd611
but I actually *want* the menu to display in such a way under 512 px. The way the code
is organized might have been the unusual element here, so I kept the order left by the
39fc76d95aa34f00d85365b3adde8ddf6efcd611

In truth, I should have converted the stylesheet to adopt 'mobile first' but couldn't
be arsed. Visualculture.less is a bit of a mess, containing 1120 lines, it's hard to
find your way around. I know why it remains as a .less file, but maybe, it could go to
a modulary .scss or .sass structure, for edition experience.