Further restructuring of the stylesheet
Now that I am more certain of my colmumn widths, I realize that they integrate
well with the majority of unit's setup by Alex for the last merge, so I'm basing
all back on those, and added a few.

Basic geometry is:
column-width = 360px
col-width = 330px
col-margin = 30px
v-unit = 20px
h-unit = 15px

The col-margin variable is in place, but I opted to use @h-unit*2 to keep things
clean.

Now I want to reorganize the chapters of the page itself, and I'll start to be
happy with it.