Structuring and commenting the stylesheet:
VISUALCULTURE LESS content map
---
The order of the stylesheet tries hard to maintain consistensy with the order
in which items appear on the web page. Below is a table of contents to help you
find your way around.

Note that since we approched a responsive website, we use a 'mobile-first'
attitude. 95% of styles will remain the same on whatever viewports, only certain
sizes change for bigger screens, so we declare all of the mobile styles first,
then we go onto a set of queries that target viewport sizes starting at 512 px
and up. This keeps the styles clean and logical, you just need to know in which
order to write things.

////////
TABLE OF CONTENT
----------------
- variables & utilities
- @font-face
- typography
- menu & navigation
- info boxes for home page
- project overview sections
  - iceberg
  - commit-list
  - commit-date
  - ...
- project detail views
- 'about' page styles
- modals
- plugins
- media queries
  - 512px & up
  - 768px & up
  - 1024px & up
  - 1280px & up